rename `LogRecords` to `logs`
rename `TapeTask` to `differentiable-Any` & rename `differentiable-float` to `differentiable-Float`& rename `differentiable-doub
le` to `differentiable-Double` & rename `differentiable-indarray` to `differentiable-INDArray`& rename `TapeTaskFactory` to `tapefactories` and ad
d `differentiable` & fix generate unidoc error & remove `Closeables`
